Move Factory's parser logic to a new dedicated Parser class
No functional or interface changes
